feat: add licence hodler spec
run some tests in parrell

add scenario data shared across tests
