+++ tests/global-setup.js
@@ -1,5 +1,5 @@
 import { request } from '@playwright/test'
-import { readFileSync } from 'fs'
+import { readFileSync, writeFileSync } from 'fs'
 
 import coreLicenceScenario from './scenarios/core-licence.js'
 
@@ -7,9 +7,11 @@ const environment = process.env.TEST_ENV ?? 'local'
 const envConfig = JSON.parse(readFileSync(`./environments/${environment}.json`, 'utf8'))
 
 export default async function globalSetup() {
+  const scenario = coreLicenceScenario()
+  writeFileSync('.scenario-data.json', JSON.stringify(scenario))
+
   const context = await request.newContext({ baseURL: envConfig.config.baseUrl })
-  // Teardown existing data before trying to add the same data again
   await context.post('/system/data/tear-down')
-  await context.post('/system/data/load', { data: coreLicenceScenario() })
+  await context.post('/system/data/load', { data: scenario })
   await context.dispose()
 }
